add `-I /path/to/clang/include` to clang args used in the `bindgen!` plugin
We do the same thing for the `bindgen` binary (see #196).
